chore: move DatabaseError and AuthError types into config/types.ts
