+++ src/lib/database-queries/comments.ts
@@ -1,7 +1,6 @@
 import { desc, eq } from "drizzle-orm";
 import { errAsync, okAsync, ResultAsync, safeTry } from "neverthrow";
-import type { CommentData } from "@/config/types";
-import type { DatabaseError } from "@/lib/database-errors";
+import type { CommentData, DatabaseError } from "@/config/types";
 import { db } from "@/lib/db/drizzle";
 import { comments } from "@/lib/db/schema";
 import { doesPostWithSlugExist } from "@/lib/dbContentQueries";
